code clean up orginze folder structure add core folder
