+++ lib/index.ts
@@ -2,4 +2,4 @@
 
 export * from "./AuthHandler";
 export * from "./types";
-export * from "./session";
+export * from "./core/session";
